requirements.txt
appdirs==1.4.4
astroid==2.4.2
click==7.1.2
colorama==0.4.3
cycler==0.10.0
distlib==0.3.0
filelock==3.0.12
Flask==1.1.2
Flask-Caching==1.9.0
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
kaleido==0.2.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
mysql-connector-python==8.0.20
numpy==1.24.4
packaging==24.1
pandas==2.0.3
plotly==5.24.1
protobuf==3.12.2
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.15.0
tenacity==9.0.0
toml==0.10.1
tzdata==2024.2
virtualenv==20.0.23
Werkzeug==1.0.1
wrapt==1.12.1
